code refectoring

+++ src/Business/AdvancedTaskInitialization.cs
@@ -24,7 +24,7 @@ namespace AdvancedTask.Business
 
         public void Initialize(InitializationEngine context)
         {
-            bool enableContentApprovalDeadline = bool.Parse(ConfigurationManager.AppSettings["ATM:EnableContentApprovalDeadline"] ?? "false");
+            var enableContentApprovalDeadline = bool.Parse(ConfigurationManager.AppSettings["ATM:EnableContentApprovalDeadline"] ?? "false");
 
             if (enableContentApprovalDeadline)
             {
@@ -32,11 +32,9 @@ namespace AdvancedTask.Business
             }
             else
             {
-                bool deleteContentApprovalDeadlineProperty = bool.Parse(ConfigurationManager.AppSettings["ATM:DeleteContentApprovalDeadlineProperty"] ?? "false");
+                var deleteContentApprovalDeadlineProperty = bool.Parse(ConfigurationManager.AppSettings["ATM:DeleteContentApprovalDeadlineProperty"] ?? "false");
                 if (deleteContentApprovalDeadlineProperty)
-                {
                     DeleteMappingProperties();
-                }
             }
 
         }
@@ -45,39 +43,33 @@ namespace AdvancedTask.Business
         {
             CreateOrDeleteTab("Content Approval", true);
 
-            foreach (ContentType contentType in _contentTypeRepository.Service.List().Where(x => x.IsAvailable))
-            {
+            foreach (var contentType in _contentTypeRepository.Service.List().Where(x => x.IsAvailable))
                 CreateUpdatePropertyDefinition(
                     contentType,
                     ContentApprovalDeadlinePropertyName,
                     "Content approval deadline",
                     typeof(PropertyDate),
                     "Content Approval", 10);
-            }
         }
 
         private void DeleteMappingProperties()
         {
-            foreach (ContentType contentType in _contentTypeRepository.Service.List().Where(x => x.IsAvailable))
-            {
+            foreach (var contentType in _contentTypeRepository.Service.List().Where(x => x.IsAvailable))
                 DeletePropertyDefinition(
                     contentType,
                     ContentApprovalDeadlinePropertyName);
-            }
         }
 
         private void CreateOrDeleteTab(string tabName, bool createNew)
         {
-            object obj2 = _lock;
+            var obj2 = _lock;
             lock (obj2)
             {
-                TabDefinition tabDefinition = this._tabDefinitionRepository.Service.Load(tabName);
+                var tabDefinition = this._tabDefinitionRepository.Service.Load(tabName);
                 if (createNew)
                 {
                     if (tabDefinition != null)
-                    {
                         return;
-                    }
 
                     tabDefinition = new TabDefinition
                     {
@@ -96,14 +88,11 @@ namespace AdvancedTask.Business
 
         private void CreateUpdatePropertyDefinition(ContentType contentType, string propertyDefinitionName, string editCaption = null, Type propertyDefinitionType = null, string tabName = null, int? propertyOrder = new int?())
         {
-            PropertyDefinition propertyDefinition =
+            var propertyDefinition =
                 GetPropertyDefinition(contentType, propertyDefinitionName);
             if (propertyDefinition == null)
             {
-                if (propertyDefinitionType == null)
-                {
-                    return;
-                }
+                if (propertyDefinitionType == null) return;
                 propertyDefinition = new PropertyDefinition();
             }
             else
@@ -113,50 +102,43 @@ namespace AdvancedTask.Business
             propertyDefinition.ContentTypeID = contentType.ID;
             propertyDefinition.DisplayEditUI = true;
             propertyDefinition.DefaultValueType = DefaultValueType.None;
+
             if (propertyDefinitionName != null)
-            {
                 propertyDefinition.Name = propertyDefinitionName;
-            }
             if (editCaption != null)
-            {
                 propertyDefinition.EditCaption = editCaption;
-            }
             if (propertyDefinitionType != null)
-            {
                 propertyDefinition.Type = this._propertyDefinitionTypeRepository.Service.Load(propertyDefinitionType);
-            }
             if (tabName != null)
             {
-                propertyDefinition.Tab = _tabDefinitionRepository.Service.Load(tabName);
+                var obj2 = _lock;
+                lock (obj2)
+                {
+                    propertyDefinition.Tab = _tabDefinitionRepository.Service.Load(tabName);
+                }
             }
+
             if (propertyOrder.HasValue)
-            {
                 propertyDefinition.FieldOrder = propertyOrder.Value;
-            }
             _propertyDefinitionRepository.Service.Save(propertyDefinition);
         }
 
         private void DeletePropertyDefinition(ContentType contentType, string propertyDefinitionName)
         {
-            PropertyDefinition propertyDefinition = GetPropertyDefinition(contentType, propertyDefinitionName);
+            var propertyDefinition = GetPropertyDefinition(contentType, propertyDefinitionName);
 
-            if (propertyDefinition != null)
-            {
-                _propertyDefinitionRepository.Service.Delete(propertyDefinition);
-            }
+            if (propertyDefinition != null) _propertyDefinitionRepository.Service.Delete(propertyDefinition);
         }
 
         private PropertyDefinition GetPropertyDefinition(ContentType contentType, string propertyName, Type propertyDefinitionType = null)
         {
-            System.Collections.Generic.IEnumerable<PropertyDefinition> source = from pd in contentType.PropertyDefinitions
-                                                                                where pd.Name.Equals(propertyName, StringComparison.OrdinalIgnoreCase)
-                                                                                select pd;
+            var source = from pd in contentType.PropertyDefinitions
+                         where pd.Name.Equals(propertyName, StringComparison.OrdinalIgnoreCase)
+                         select pd;
             if (propertyDefinitionType != null)
-            {
                 source = from pd in source
                          where pd.Type.DefinitionType == propertyDefinitionType
                          select pd;
-            }
             return source.FirstOrDefault();
         }
